refact js para chart
